+++ lib/tool_schema_generator.dart
@@ -8,3 +8,6 @@
 library tool_schema_generator;
 
 export 'src/annotations.dart';
+export 'src/tool_exceptions.dart';
+export 'src/tool_registry.dart';
+export 'src/tool_result.dart';